feat(contracts): update TrialMarket contract

+++ packages/contracts/contracts/TrialMarket.sol
@@ -133,5 +133,5 @@ contract TrialMarket is Ownable, ReentracyGuard, FunctionsClient, AutomationComp
      * theoretically stake on both sides (hedging).
      */
     mapping(uint256 => mapping(address => uint256)) public yesPositions;
-    mapping(uint256 => mapping(address => uint))
+    mapping(uint256 => mapping(address => uint256))
 }
